requirements.txt
numpy==2.2.6
pandas==2.3.0
bottleneck==1.5.0
numexpr==2.11.0
# 指标库
ft-pandas-ta==0.3.15
ta-lib==0.5.5
technical==1.5.1

ccxt==4.4.91
cryptography==45.0.4
aiohttp==3.12.13
SQLAlchemy==2.0.41
python-telegram-bot==22.2
# 无法硬固定版本，因为telegram-bot用~符号固定了httpx版本
httpx>=0.24.1
humanize==4.12.3
cachetools==6.1.0
requests==2.32.4
urllib3==2.5.0
certifi==2025.6.15
jsonschema==4.24.0
tabulate==0.9.0
pycoingecko==3.2.0
jinja2==3.1.6
joblib==1.5.1
rich==14.0.0
pyarrow==20.0.0; platform_machine != 'armv7l'

# 加快30%的ticker文件加载速度
python-rapidjson==1.20
# 正确格式化API响应
orjson==3.10.18

# 通知systemd系统
sdnotify==0.3.2

# API服务器
fastapi==0.115.14
pydantic==2.11.7
uvicorn==0.35.0
pyjwt==2.10.1
aiofiles==24.1.0
psutil==7.0.0

# 交互式构建配置文件
questionary==2.1.0
prompt-toolkit==3.0.51
# datetime库的扩展
python-dateutil==2.9.0.post0
pytz==2025.2

# 定时任务
schedule==1.2.2

# WebSocket消息
websockets==15.0.1
janus==2.0.0

ast-comments==1.2.3
packaging==25.0